+++ example/lib/pages/home.dart
@@ -144,7 +144,7 @@ class Home extends StatelessWidget {
                 Get.toNamed("/betterSwipeAction");
               },
             ),
-            // SizedBox(height: 20.bw),
+            SizedBox(height: 20.bw),
             // Text("展示组件".tr, style: TextStyle(fontSize: 14.bsp)),
             // SizedBox(height: 10.bw),
             // BetterCell(